+++ README.md
@@ -1,16 +1,5 @@
 # neighborhood-bake-sale
 
-Commands:
-```
-$ npm start
-```
-Use npm start to run the server, it will auto refresh react as well
-
-```
-$ npm run build
-```
-When we are ready to deploy, use npm run build to create a build folder. The actual backend server will still need to be ran with nodemon server.js
-
 # For reference only, please update this later
 Stripe js (node)
 https://github.com/stripe/stripe-node
