+++ app/components/Seller/StoreDescription.js
@@ -1,0 +1,27 @@
+
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+
+class StoreDescription extends Component {
+    constructor(props) {
+      super(props);
+      this.state = {
+        
+      };
+  
+    }
+  
+    render() {
+      return (
+         <div>  
+           <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse aliquam dui id mauris pharetra luctus. 
+               Sed at nisl in dolor vulputate mollis eu et odio. Suspendisse malesuada, lorem vitae bibendum euismod, arcu turpis eleifend arcu, dignissim aliquam ligula nunc id sapien. 
+               Nullam vel blandit erat. Morbi velit libero, vestibulum in ipsum at, accumsan pretium mauris. Sed risus turpis, volutpat nec euismod ac, tempus at orci. 
+               Nulla egestas lectus non turpis molestie, nec pharetra turpis scelerisque. Curabitur ac accumsan dolor.</p>
+         </div>
+      );
+    }
+  }
+  
+  export default StoreDescription;
+
