+++ app/components/Seller/StoreHours.js
@@ -1,0 +1,24 @@
+
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+
+class StoreHours extends Component {
+    constructor(props) {
+      super(props);
+      this.state = {
+        
+      };
+  
+    }
+  
+    render() {
+      return (
+         <div>  
+           <h1>4:00PM-5:00PM</h1>
+         </div>
+      );
+    }
+  }
+  
+  export default StoreHours;
+
