+++ app/components/Seller/StoreTitle.js
@@ -1,0 +1,25 @@
+
+import React, { Component } from 'react';
+import ReactDOM from 'react-dom';
+import { BrowserRouter as Router, Route, Link } from 'react-router-dom';
+
+class StoreTitle extends Component {
+    constructor(props) {
+      super(props);
+      this.state = {
+        
+      };
+  
+    }
+  
+    render() {
+      return (
+         <div>  
+           <h1>Store Title Goes Here</h1>
+         </div>
+      );
+    }
+  }
+  
+  export default StoreTitle;
+
