+++ app/components/children/SearchMap.js
@@ -1,0 +1,1 @@
+// Put stuff here
