+++ app/components/children/Store.js
@@ -1,0 +1,1 @@
+// Put stuff here
