+++ app/index.css
@@ -1,0 +1,15 @@
+body {
+    background: whitesmoke;
+}
+/* .nav {
+    display: flex;
+}
+
+.active {
+    font-weight: bold;
+}
+
+.nav {
+    margin-right: 15px;
+    margin-left: 15px;
+} */
