+++ app/index.html
@@ -15,5 +15,5 @@
         <div id="app"></div>
 
         <!-- All of our JS code will be transpiled into this bundle.js file -->
-    <script type="text/javascript" src="bundle.js"></script></body>
+    </body>
 </html>
