--- src/components/Seller/Menu.js
@@ -1,37 +1,0 @@
-
-import React, { Component } from 'react';
-import MenuItem from "./MenuItem";
-
-class Menu extends Component {
-    constructor(props) {
-      super(props);
-      this.state = {
-        
-      };
-  
-    }
-
-    createMenuItems() {
-      let menuList;
-      menuList = this.props.menu.map((menuItem, i)=>{
-        return <MenuItem key={i} index={i} item={menuItem} edit={this.props.edit} updateState={this.props.updateState}/>
-      });
-      return menuList;
-    }
-  
-    render() {
-      if(this.props.menu.length > 1){
-        return (
-          <ul>  
-            {this.createMenuItems()}
-          </ul>
-       );
-      } 
-      return (
-        <h4>No Items</h4>
-      );
-    }
-  }
-  
-  export default Menu;
-
