--- src/components/Store/WriteReview/Rating.js
@@ -1,17 +1,0 @@
-import React, {Component} from 'react';
-
-class Rating extends Component {
-  render() {
-    return (
-      <div className="rating">
-        <span>☆</span>
-        <span>☆</span>
-        <span>☆</span>
-        <span>☆</span>
-        <span>☆</span>
-      </div>
-    )
-  }
-}
-
-export default Rating;
