--- src/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './style.css';
-import Main from './components/Main';
-import registerServiceWorker from './registerServiceWorker';
-
-ReactDOM.render(<Main />, document.getElementById('root'));
-registerServiceWorker();
